# BACKEND/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/verify-email": {
      "post": {
        "operationId": "NodeController_verifyEmail",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyEmailDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Verify email — trigger OTP (public)",
        "tags": [
          "Auth & Node"
        ]
      }
    },
    "/auth/reset-password": {
      "post": {
        "operationId": "NodeController_resetPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Reset password using OTP (public)",
        "tags": [
          "Auth & Node"
        ]
      }
    },
    "/auth/warehouse": {
      "post": {
        "operationId": "NodeController_GetCredentials1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Warehouse login (public)",
        "tags": [
          "Auth & Node"
        ]
      }
    },
    "/auth/superuser": {
      "post": {
        "operationId": "NodeController_GetCredentialsSU",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Super User login (public)",
        "tags": [
          "Auth & Node"
        ]
      }
    },
    "/auth/transit": {
      "post": {
        "operationId": "NodeController_GetCredentials2",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Transit Hub login (public)",
        "tags": [
          "Auth & Node"
        ]
      }
    },
    "/auth/agency": {
      "post": {
        "operationId": "NodeController_GetCredential3",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Local Agency login (public)",
        "tags": [
          "Auth & Node"
        ]
      }
    },
    "/auth/profile/{id}": {
      "post": {
        "operationId": "NodeController_updateProfile",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Node ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Update node profile (public)",
        "tags": [
          "Auth & Node"
        ]
      }
    },
    "/auth/password/{id}": {
      "post": {
        "operationId": "NodeController_updatePassword",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Node ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePasswordDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Update node password (public)",
        "tags": [
          "Auth & Node"
        ]
      }
    },
...
    "/{warehouseId}/stockInventory": {
      "get": {
        "operationId": "NodeController_getCurentStock",
        "parameters": [
          {
            "name": "warehouseId",
            "required": true,
            "in": "path",
            "description": "Warehouse ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get current stock inventory for a warehouse",
        "tags": [
          "Auth & Node"
        ]
      }
    },
    "/{warehouseId}/stockAddition": {
      "post": {
        "operationId": "NodeController_addStockItem",
        "parameters": [
          {
            "name": "warehouseId",
            "required": true,
            "in": "path",
            "description": "Warehouse ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StockAdditionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Add a stock item to warehouse inventory",
        "tags": [
          "Auth & Node"
        ]
      }
    },
    "/{warehouseId}/pendingShipments": {
      "get": {
        "operationId": "NodeController_getPendingShipments",
        "parameters": [
          {
            "name": "warehouseId",
            "required": true,
            "in": "path",
            "description": "Warehouse ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get pending shipments for a warehouse",
        "tags": [
          "Auth & Node"
        ]
      },
      "post": {
        "operationId": "NodeController_addPendingShipment",
        "parameters": [
          {
            "name": "warehouseId",
            "required": true,
            "in": "path",
            "description": "Warehouse ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePendingShipmentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Add a pending shipment to a warehouse",
        "tags": [
          "Auth & Node"
        ]
      }
    },
    "/{warehouseId}/pendingShipments/{orderId}": {
      "delete": {
        "operationId": "NodeController_deletePendingShipment",
        "parameters": [
          {
            "name": "warehouseId",
            "required": true,
            "in": "path",
            "description": "Warehouse ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "orderId",
            "required": true,
            "in": "path",
            "description": "Order ID to delete",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Delete a pending shipment from a warehouse",
        "tags": [
          "Auth & Node"
        ]
      }
    },
    "/hub/{hubId}/inventory": {
      "get": {
        "operationId": "NodeController_getHubInventory",
        "parameters": [
          {
            "name": "hubId",
            "required": true,
            "in": "path",
            "description": "Transit Hub ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get inventory at a transit hub",
        "tags": [
          "Auth & Node"
        ]
      }
    },
    "/hub/{hubId}/capacity": {
      "get": {
        "operationId": "NodeController_getHubCapacity",
        "parameters": [
          {
            "name": "hubId",
            "required": true,
            "in": "path",
            "description": "Transit Hub ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get capacity information for a transit hub",
        "tags": [
          "Auth & Node"
        ]
      }
    },
    "/agency/{agencyId}/shipments": {
      "get": {
        "operationId": "NodeController_getAgencyShipments",
        "parameters": [
          {
            "name": "agencyId",
            "required": true,
            "in": "path",
            "description": "Agency ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get shipments assigned to an agency",
        "tags": [
          "Auth & Node"
        ]
      }
    },
...
  },
  "info": {
    "title": "WARE2DOOR API",
    "description": "## WARE2DOOR Logistics API\n\nEnd-to-end logistics platform covering warehouse management, transit hub scanning, last-mile delivery, RTO lifecycle, and super-user administration.\n\n### Authentication\nProtected endpoints require an **`x-role`** header specifying the caller's role.\n\nClick the **Authorize 🔒** button, then enter one of the allowed values:\n\n| Role Value | Access |\n|---|---|\n| `WAREHOUSE` | Warehouse management endpoints |\n| `TRANSIT_HUB` | Transit hub scan & reports endpoints |\n| `LOCAL_AGENCY` | Agency delivery & agent endpoints |\n| `SUPER_USER` | Full administrative access |\n\n> **Public endpoints** (login, verify-email, track, raise-ticket) do **not** require this header.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "x-role": {
        "type": "apiKey",
        "in": "header",
        "name": "x-role",
        "description": "Role-based access header. Allowed values: WAREHOUSE | TRANSIT_HUB | LOCAL_AGENCY | SUPER_USER"
      }
    },
    "schemas": {
...
      "VerifyEmailDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@ware2door.com",
            "description": "Email to verify / send OTP to"
          }
        },
        "required": [
          "email"
        ]
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@ware2door.com",
            "description": "Email address to reset password for"
          },
          "otp": {
            "type": "string",
            "example": "000000",
            "description": "6-digit OTP (use 000000 for now)"
          },
          "newPassword": {
            "type": "string",
            "example": "newSecurePass123",
            "description": "New password (min 6 characters)"
          }
        },
        "required": [
          "email",
          "otp",
          "newPassword"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "admin@ware2door.com",
            "description": "Registered email address"
          },
          "password": {
            "type": "string",
            "example": "password123",
            "description": "Account password"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "UpdateProfileDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "John Doe",
            "description": "Full name of the user"
          },
          "email": {
            "type": "string",
            "example": "john@ware2door.com",
            "description": "New email address"
          },
          "phone": {
            "type": "string",
            "example": "+91 9876543210",
            "description": "Valid 10-digit mobile number with optional country code"
          }
        }
      },
      "UpdatePasswordDto": {
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string",
            "example": "oldPassword123",
            "description": "Current account password"
          },
          "newPassword": {
            "type": "string",
            "example": "newPassword456",
            "description": "New password (min 6 characters)"
          }
        },
        "required": [
          "currentPassword",
          "newPassword"
        ]
      },
      "StockAdditionDto": {
        "type": "object",
        "properties": {
          "labelID": {
            "type": "string",
            "example": "LBL-001",
            "description": "Optional label ID for the stock item"
          },
          "itemName": {
            "type": "string",
            "example": "Laptop",
            "description": "Name of the item being added to stock"
          },
          "quantity": {
            "type": "number",
            "example": 10,
            "description": "Quantity to add (minimum 1)"
          }
        },
        "required": [
          "itemName",
          "quantity"
        ]
      },
      "CreatePendingShipmentDto": {
        "type": "object",
        "properties": {
          "orderId": {
            "type": "string",
            "example": "ORD-12345",
            "description": "Unique order ID"
          },
          "source": {
            "type": "string",
            "example": "WH-HYD-001",
            "description": "Source warehouse ID"
          },
          "customerName": {
            "type": "string",
            "example": "Ravi Kumar",
            "description": "Customer full name"
          },
          "customerPhone": {
            "type": "string",
            "example": "9876543210",
            "description": "Customer phone (10-digit, optional country code)"
          },
          "customerAddress": {
            "type": "string",
            "example": "12, MG Road, Bangalore - 560001",
            "description": "Full delivery address"
          },
          "productName": {
            "type": "string",
            "example": "Wireless Mouse",
            "description": "Name of the product being shipped"
          },
          "sku": {
            "type": "string",
            "example": "SKU-WM-BLK",
            "description": "Product SKU code"
          },
          "qty": {
            "type": "number",
            "example": 2,
            "description": "Quantity of items (minimum 1)"
          },
          "priority": {
            "type": "string",
            "example": "HIGH",
            "enum": [
              "LOW",
              "NORMAL",
              "HIGH"
            ],
            "description": "Shipment priority level"
          }
        },
        "required": [
          "orderId",
          "source",
          "customerName",
          "customerPhone",
          "customerAddress",
          "productName",
          "qty"
        ]
      },
...
    }
  },
  "security": [
    {
      "x-role": []
    }
  ]
}